Fix Banned User menu overflow
