+++ app/javascript/components/admin/manage_users/BannedPendingRow.jsx
@@ -1,13 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {
-  Stack, Navbar, Container,
+  Stack, Dropdown,
 } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
+import {
+  CheckCircleIcon,
+  CheckIcon,
+  EllipsisVerticalIcon,
+  XCircleIcon,
+} from '@heroicons/react/24/outline';
 import Avatar from '../../users/user/Avatar';
+import useUpdateUserStatus from '../../../hooks/mutations/admin/manage_users/useUpdateUserStatus';
 
-export default function BannedPendingRow({ user, children }) {
+export default function BannedPendingRow({ user, pendingTable }) {
   const { t } = useTranslation();
+  const updateUserStatus = useUpdateUserStatus();
+
   return (
     <tr key={user.id} className="align-middle text-muted border border-2">
       <td className="border-end-0">
@@ -25,13 +34,28 @@ export default function BannedPendingRow({ user, children }) {
       <td className="border-0"> {user.email} </td>
       <td className="border-0"> {user.created_at} </td>
       <td className="border-start-0">
-        <Navbar>
-          <Container>
-            <div className="d-inline-flex">
-              {children}
-            </div>
-          </Container>
-        </Navbar>
+        <Dropdown className="float-end cursor-pointer">
+          <Dropdown.Toggle className="hi-s" as={EllipsisVerticalIcon} />
+          <Dropdown.Menu>
+            {pendingTable ? (
+              <>
+                <Dropdown.Item onClick={() => updateUserStatus.mutate({ id: user.id, status: 'active' })}>
+                  <CheckCircleIcon className="hi-s me-2" />
+                  {t('admin.manage_users.approve')}
+                </Dropdown.Item>
+                <Dropdown.Item onClick={() => updateUserStatus.mutate({ id: user.id, status: 'banned' })}>
+                  <XCircleIcon className="hi-s me-2" />
+                  {t('admin.manage_users.decline')}
+                </Dropdown.Item>
+              </>
+            ) : (
+              <Dropdown.Item onClick={() => updateUserStatus.mutate({ id: user.id, status: 'active' })}>
+                <CheckIcon className="hi-s me-2" />
+                {t('admin.manage_users.unban')}
+              </Dropdown.Item>
+            )}
+          </Dropdown.Menu>
+        </Dropdown>
       </td>
     </tr>
   );
@@ -45,5 +69,5 @@ BannedPendingRow.propTypes = {
     email: PropTypes.string.isRequired,
     created_at: PropTypes.string.isRequired,
   }).isRequired,
-  children: PropTypes.node.isRequired,
+  pendingTable: PropTypes.bool.isRequired,
 };
